+++ src/interaction/random_options_explorer.py
@@ -1,26 +1,12 @@
 """An explorer that takes random options."""
 
-from typing import List, Set
-
-from gym.spaces import Box
-
-from predicators.src.approaches.random_options_approach import \
-    RandomOptionsApproach
+from predicators.src import utils
 from predicators.src.interaction import BaseExplorer
-from predicators.src.structs import ExplorationStrategy, ParameterizedOption, \
-    Predicate, Task, Type
+from predicators.src.structs import ExplorationStrategy, Task
 
 
 class RandomOptionsExplorer(BaseExplorer):
     """Samples random options."""
-
-    def __init__(self, predicates: Set[Predicate],
-                 options: Set[ParameterizedOption], types: Set[Type],
-                 action_space: Box, train_tasks: List[Task]) -> None:
-        super().__init__(predicates, options, types, action_space, train_tasks)
-        # Set up a random options approach.
-        self._random_options_approach = RandomOptionsApproach(
-            predicates, options, types, action_space, train_tasks)
 
     @classmethod
     def get_name(cls) -> str:
@@ -29,7 +15,9 @@ class RandomOptionsExplorer(BaseExplorer):
     def get_exploration_strategy(self, task: Task,
                                  timeout: int) -> ExplorationStrategy:
         # Take random options.
-        policy = self._random_options_approach.solve(task, timeout)
+        policy = utils.create_random_option_policy(self._options,
+                                                   self._action_space,
+                                                   self._rng)
         # Never terminate (until the interaction budget is exceeded).
         termination_function = lambda _: False
         return policy, termination_function
